Broken state, pausing due to headache... Breaking up projectiles and explosions in the same manner as enemies. Moved some general functionality to Entity TU. Moved many definitions to _types.cpp to decouple specific from general functionality. Added TU for each projectile type. Removed advanced bullet, laser, and missile types and added tiers to bullet, laser, and missile TUs.

+++ src/main/system/onegungame.cpp
@@ -138,7 +138,7 @@ namespace OneGunGame {
             [](auto entity, Renderable &renderable, Rotating &rotating) {
                 spdlog::trace("Updating entity {}: Rotating {} radians per second", 
                 static_cast<int>(entity), rotating.RotationsPerSecond.asRadians());
-                renderable.Sprite.rotate(rotating.RotationsPerSecond / Configuration.FrameRateLimit);
+                renderable.Sprite.rotate(rotating.RotationsPerSecond / static_cast<float>(Configuration.FrameRateLimit));
             }
         );
